backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
pymongo==4.6.1
bcrypt==4.1.2
eventlet==0.35.2

# AI/ML Dependencies
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Additional utilities
faker==20.1.0
python-dateutil==2.8.2

